more direction cleanup, spec helper cleanup, renamed matchers
