+++ lib/kitchen/directions/move_title_text_into_span.rb
@@ -1,13 +1,15 @@
 module Kitchen
   module Directions
+    module MoveTitleTextIntoSpan
 
-    def self.move_title_text_into_span(title)
-      title.replace_children(with:
-        <<~HTML
-          <span data-type="" itemprop="" class="os-text">#{title.text}</span>
-        HTML
-      )
+      def self.v1(title:)
+        title.replace_children(with:
+          <<~HTML
+            <span data-type="" itemprop="" class="os-text">#{title.text}</span>
+          HTML
+        )
+      end
+
     end
-
   end
 end
